Ensure script has absolute path
I was doing setup on a new machine after some time and when installing
the script service I was in the script service folder which meant that
when I tab completed the argument ./service_script.py was passed to the
CLI which passed it to the bin path for the sc.exe install and of course
none of that worked. So I am going to abuse the type field of the
in add arguement to have it run the input through abspath to ensure we
have the full qualified path to the script.